+++ bin/faucet/src/server/resources/index.js
@@ -6,7 +6,7 @@ class MidenFaucet {
         this.publicButton = document.getElementById('send-public-button');
         this.faucetAddress = document.getElementById('faucet-address');
         this.progressFill = document.getElementById('progress-fill');
-        this.tokensClaimed = document.getElementById('tokens-claimed');
+        this.issuance = document.getElementById('issuance');
         this.tokensSupply = document.getElementById('tokens-supply');
 
         // Check if SHA3 is available right from the start
@@ -57,7 +57,6 @@ class MidenFaucet {
 
         try {
             await this.getTokens(powData.challenge, nonce, recipient, amount, isPrivateNote);
-            this.resetForm();
         } catch (error) {
             this.showError(`Failed to send tokens: ${error.message}`);
         }
@@ -74,10 +73,10 @@ class MidenFaucet {
                     option.textContent = amount;
                     this.tokenSelect.appendChild(option);
                 }
-                // TODO: add metadata values
-                this.tokensClaimed.textContent = '0';
-                this.tokensSupply.textContent = '100,000,000,000';
-                this.progressFill.style.width = '0%';
+
+                this.issuance.textContent = data.issuance.toLocaleString();
+                this.tokensSupply.textContent = data.max_supply.toLocaleString();
+                this.progressFill.style.width = (data.issuance / data.max_supply) * 100 + '%';
             })
             .catch(error => {
                 console.error('Error fetching metadata:', error);